feat: add test and refactored main
