+++ src/index.html
@@ -6,7 +6,7 @@
 		</head>
 		<body>
 			<h1 class="pageTitle">Title for page</h1>
-			<span>Hello, new 2</span>
+			<span>Hello, wefri</span>
 			<ul>
 				
 					
@@ -26,7 +26,7 @@
 					
 				
 			</ul>
-			<a href="https://github.com/Agile-Arch-Angels/decentralized-tech-radar_dev/" title="Tech radar" ref="rel">Tech radar repo</a>
+			<a href="//github.com/Agile-Arch-Angels/decentralized-tech-radar_dev/" title="Tech radar" ref="rel">Tech radar repo</a>
 		</body>
 	</html>
